# docs/swagger.yaml
basePath: /
definitions:
...
  dto.CreateOrder:
    properties:
      payment_id:
        type: integer
      schedule_id:
        description: Id           string  `json:"order_id,omitempty"`
        type: integer
      seat_id:
        items:
          type: integer
        type: array
      sub_total:
        type: number
      tax:
        type: number
      total_price:
        type: number
      total_ticket:
        type: integer
    type: object
...
  dto.History:
    properties:
      booking_code:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      movie_title:
        type: string
      order_id:
        type: string
      payment_method:
        type: string
      phone:
        type: string
      seat_number:
        type: string
      show_date:
        type: string
      show_time:
        type: string
      status:
        type: string
      total_price:
        type: string
    type: object
...
  dto.MovieDetails:
    properties:
      actor_name:
        type: string
      background_url:
        type: string
      director_name:
        type: string
      duration:
        type: integer
      genres_name:
        type: string
      id_movie:
        type: integer
      poster_url:
        type: string
      release_date:
        type: string
      synopsis:
        type: string
      title:
        type: string
    type: object
  dto.MovieGenres:
    properties:
      id:
        type: integer
      name:
        type: string
      title:
        type: string
    type: object
  dto.Movies:
    properties:
      Id:
        type: integer
      background_url:
        type: string
      duration:
        type: integer
      poster_url:
        type: string
      rating:
        type: number
      release_date:
        type: string
      status:
        type: string
      synopsis:
        type: string
      title:
        type: string
    required:
    - Id
    type: object
...
  dto.ResponseError:
    properties:
      error:
        type: string
      msg:
        type: string
      success:
        type: boolean
    type: object
  dto.Schedules:
    properties:
      cinema_address:
        type: string
      cinema_city:
        type: string
      cinema_name:
        type: string
      movie_id:
        type: integer
      movie_name:
        type: string
      price:
        type: integer
      show_date:
        type: string
      show_time:
        type: string
      studio_name:
        type: string
    required:
    - cinema_address
    - cinema_city
    - cinema_name
    - movie_id
    - movie_name
    - price
    - show_date
    - show_time
    - studio_name
    type: object
  dto.SearchMovies:
    properties:
      genre_name:
        type: string
      id:
        type: integer
      poster_url:
        type: string
      title:
        type: string
    type: object
...
  dto.Users:
    properties:
      email:
        type: string
      first_name:
        type: string
      id:
        type: integer
      image:
        type: string
      last_name:
        type: string
      phone:
        type: string
    type: object
host: localhost:8002
info:
  contact: {}
  description: Backend Services For Tickitz Project
  title: Tickitz Services
  version: "1.0"
paths:
...
  /movies/:
    get:
      parameters:
      - description: Status Movies (popular, upcoming or now_showing)
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Movies'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ResponseError'
      summary: Get Movies By Status
      tags:
      - movies
  /movies/{id}:
    get:
      parameters:
      - description: Id Movie
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.MovieDetails'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ResponseError'
      summary: Get Movies Detail By Id
      tags:
      - movies
  /movies/genres/{id}:
    get:
      parameters:
      - description: Id Movie
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.MovieGenres'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ResponseError'
      summary: Get Movie Genres By Id
      tags:
      - movies
  /movies/search:
    get:
      parameters:
      - description: Movie Title
        in: query
        name: title
        type: string
      - collectionFormat: multi
        description: Genres
        in: query
        items:
          type: string
        name: genre
        type: array
      - description: Page
        in: query
        name: page
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SearchMovies'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ResponseError'
      summary: Search Movies
      tags:
      - movies
  /order/:
    post:
      consumes:
      - application/json
      parameters:
      - description: Create Order Body
        in: body
        name: auth
        required: true
        schema:
          $ref: '#/definitions/dto.CreateOrder'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.CreateOrder'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ResponseError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ResponseError'
      security:
      - BearerAuth: []
      summary: Create Order
      tags:
      - order
  /order/schedule:
    get:
      parameters:
      - description: Movie Date
        in: query
        name: date
        required: true
        type: string
      - description: Movie Time
        in: query
        name: time
        required: true
        type: string
      - description: Cinema City
        in: query
        name: location
        required: true
        type: string
      - description: Movie Id
        in: query
        name: movie_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Schedules'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ResponseError'
      security:
      - BearerAuth: []
      summary: Get Schedules
      tags:
      - order
  /user/:
    patch:
      consumes:
      - application/json
      parameters:
      - description: First Name
        in: formData
        name: first_name
        type: string
      - description: Last Name
        in: formData
        name: last_name
        type: string
      - description: Update Profile Body
        in: formData
        name: image
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.MovieGenres'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ResponseError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ResponseError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ResponseError'
      security:
      - BearerAuth: []
      summary: Update Profile
      tags:
      - user
  /user/history:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.History'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ResponseError'
      security:
      - BearerAuth: []
      summary: Get User History
      tags:
      - user
  /user/profile:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Users'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ResponseError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ResponseError'
      security:
      - BearerAuth: []
      summary: Get User Profiles By Email
      tags:
      - user
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by space and JWT Token
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"